refactor: validateResponse 위치 이동
